+++ lib/src/televerse/models/models.dart
@@ -14,3 +14,4 @@ part 'http_exception.dart';
 part 'long_polling_exception.dart';
 part 'televerse_exception.dart';
 part 'mention.dart';
+part 'custom_emoji.dart';